drivers: fix type for limit variable

+++ src/drivers/cpu/driver_cpu.c
@@ -191,8 +191,8 @@ _starpu_get_worker_from_driver(struct starpu_driver *d)
 
 static size_t _starpu_cpu_get_global_mem_size(int devid, struct _starpu_machine_config *config)
 {
-	ssize_t global_mem;
-	ssize_t limit;
+	size_t global_mem;
+	int limit;
 
 	limit = starpu_get_env_number("STARPU_LIMIT_CPU_MEM");
 #ifdef STARPU_DEVEL
